[validation] Move slug validation from @sanity/schema to @sanity/validation
